Playing with seperating widgets
